# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.1.0
  description: GAuth Management API — RFC 0118 §11.1 mandate lifecycle management
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /gauth/mgmt/v1/profiles:
    get:
      operationId: listProfiles
      tags: [profiles]
      summary: List governance profiles
      description: Returns all available governance profiles with descriptions
      security:
        - callerIdentity: []
      responses:
        "200":
          description: List of governance profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProfileInfo"
  /gauth/mgmt/v1/profiles/{profile}/ceilings:
    get:
      operationId: getProfileCeilings
      tags: [profiles]
      summary: Get profile ceilings
      description: Returns the ceiling table for a specific governance profile
      security:
        - callerIdentity: []
      parameters:
        - name: profile
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/GovernanceProfile"
      responses:
        "200":
          description: Profile ceiling table
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileCeilings"
        "404":
          $ref: "#/components/responses/ManagementError"
  /gauth/mgmt/v1/mandates:
    post:
      operationId: createMandate
      tags: [mandates]
      summary: Create a DRAFT mandate
      description: Creates a new mandate in DRAFT status after passing 3-stage validation
      security:
        - callerIdentity: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MandateCreationRequest"
      responses:
        "201":
          description: Mandate created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreationResponse"
        "403":
          $ref: "#/components/responses/ManagementError"
        "422":
          $ref: "#/components/responses/ManagementError"
    get:
      operationId: listMandates
      tags: [mandates]
      summary: List mandates
      description: List mandates with optional filtering and cursor-based pagination
      security:
        - callerIdentity: []
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/MandateStatus"
        - name: agent_id
          in: query
          schema:
            type: string
        - name: project_id
          in: query
          schema:
            type: string
        - name: governance_profile
          in: query
          schema:
            $ref: "#/components/schemas/GovernanceProfile"
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Paginated mandate list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedMandateResponse"
        "403":
          $ref: "#/components/responses/ManagementError"
  /gauth/mgmt/v1/mandates/{id}:
    get:
      operationId: getMandate
      tags: [mandates]
      summary: Get mandate detail
      description: Returns full mandate detail with computed budget utilization
      security:
        - callerIdentity: []
      parameters:
        - $ref: "#/components/parameters/MandateId"
      responses:
        "200":
          description: Mandate detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MandateResponse"
        "404":
          $ref: "#/components/responses/ManagementError"
    delete:
      operationId: deleteMandate
      tags: [mandates]
      summary: Delete a DRAFT mandate
      description: Soft-deletes a DRAFT mandate (sets status to DELETED)
      security:
        - callerIdentity: []
      parameters:
        - $ref: "#/components/parameters/MandateId"
      responses:
        "200":
          description: Mandate deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
        "404":
          $ref: "#/components/responses/ManagementError"
        "409":
          $ref: "#/components/responses/ManagementError"
  /gauth/mgmt/v1/mandates/{id}/history:
    get:
      operationId: getMandateHistory
      tags: [mandates]
      summary: Get mandate audit history
      description: Returns the full immutable audit trail for a mandate
      security:
        - callerIdentity: []
      parameters:
        - $ref: "#/components/parameters/MandateId"
      responses:
        "200":
          description: Audit log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditLogEntry"
        "404":
          $ref: "#/components/responses/ManagementError"
  /gauth/mgmt/v1/mandates/{id}/activate:
    post:
      operationId: activateMandate
      tags: [mandates]
      summary: Activate a DRAFT mandate
      description: Transitions DRAFT→ACTIVE with re-validation and atomic supersession
      security:
        - callerIdentity: []
      parameters:
        - $ref: "#/components/parameters/MandateId"
      responses:
        "200":
          description: Mandate activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivationResponse"
        "404":
          $ref: "#/components/responses/ManagementError"
        "409":
          $ref: "#/components/responses/ManagementError"
        "422":
          $ref: "#/components/responses/ManagementError"
  /gauth/mgmt/v1/mandates/{id}/revoke:
    post:
      operationId: revokeMandate
      tags: [mandates]
      summary: Revoke a mandate
      description: Transitions ACTIVE|SUSPENDED→REVOKED with cascade to child delegations
      security:
        - callerIdentity: []
      parameters:
        - $ref: "#/components/parameters/MandateId"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  default: ""
      responses:
        "200":
          description: Mandate revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevocationResponse"
        "404":
          $ref: "#/components/responses/ManagementError"
        "409":
          $ref: "#/components/responses/ManagementError"
  /gauth/mgmt/v1/mandates/{id}/suspend:
    post:
      operationId: suspendMandate
      tags: [mandates]
      summary: Suspend a mandate
      description: Transitions ACTIVE→SUSPENDED with cascade to child delegations
      security:
        - callerIdentity: []
      parameters:
        - $ref: "#/components/parameters/MandateId"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  default: ""
      responses:
        "200":
          description: Mandate suspended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuspensionResponse"
        "404":
          $ref: "#/components/responses/ManagementError"
        "409":
          $ref: "#/components/responses/ManagementError"
  /gauth/mgmt/v1/mandates/{id}/resume:
    post:
      operationId: resumeMandate
      tags: [mandates]
      summary: Resume a suspended mandate
      description: Transitions SUSPENDED→ACTIVE with expiry check and cascade resume
      security:
        - callerIdentity: []
      parameters:
        - $ref: "#/components/parameters/MandateId"
      responses:
        "200":
          description: Mandate resumed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumptionResponse"
        "404":
          $ref: "#/components/responses/ManagementError"
        "409":
          $ref: "#/components/responses/ManagementError"
  /gauth/mgmt/v1/mandates/{id}/budget/increase:
    post:
      operationId: increaseBudget
      tags: [budget]
      summary: Increase mandate budget
      description: Additive-only budget increase on ACTIVE or SUSPENDED mandates
      security:
        - callerIdentity: []
      parameters:
        - $ref: "#/components/parameters/MandateId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - additional_cents
              properties:
                additional_cents:
                  type: integer
                  minimum: 1
      responses:
        "200":
          description: Budget increased
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetIncreaseResponse"
        "404":
          $ref: "#/components/responses/ManagementError"
        "409":
          $ref: "#/components/responses/ManagementError"
  /gauth/mgmt/v1/mandates/{id}/budget/consume:
    post:
      operationId: consumeBudget
      tags: [budget]
      summary: Report budget consumption
      description: Idempotent PEP consumption report deduplicated by mandate_id + enforcement_request_id
      security:
        - callerIdentity: []
      parameters:
        - $ref: "#/components/parameters/MandateId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - enforcement_request_id
                - amount_cents
              properties:
                enforcement_request_id:
                  type: string
                  minLength: 1
                amount_cents:
                  type: integer
                  minimum: 1
                description:
                  type: string
                  default: ""
      responses:
        "200":
          description: Consumption recorded (or idempotent acknowledgement)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsumptionResponse"
        "404":
          $ref: "#/components/responses/ManagementError"
        "409":
          $ref: "#/components/responses/ManagementError"
        "422":
          $ref: "#/components/responses/ManagementError"
  /gauth/mgmt/v1/mandates/{id}/budget:
    get:
      operationId: getBudget
      tags: [budget]
      summary: Get budget state
      description: Returns current budget utilization for a mandate
      security:
        - callerIdentity: []
      parameters:
        - $ref: "#/components/parameters/MandateId"
      responses:
        "200":
          description: Budget state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetDetail"
        "404":
          $ref: "#/components/responses/ManagementError"
  /gauth/mgmt/v1/mandates/{id}/ttl/extend:
    post:
      operationId: extendTtl
      tags: [ttl]
      summary: Extend mandate TTL
      description: Additive-only TTL extension validated against governance profile ceiling
      security:
        - callerIdentity: []
      parameters:
        - $ref: "#/components/parameters/MandateId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - additional_seconds
              properties:
                additional_seconds:
                  type: integer
                  minimum: 1
      responses:
        "200":
          description: TTL extended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TTLExtensionResponse"
        "404":
          $ref: "#/components/responses/ManagementError"
        "409":
          $ref: "#/components/responses/ManagementError"
        "422":
          $ref: "#/components/responses/ManagementError"
  /gauth/mgmt/v1/delegations:
    post:
      operationId: createDelegation
      tags: [delegations]
      summary: Create a child delegation
      description: Creates child mandate with scope narrowing, budget deduction, depth check, TTL subset
      security:
        - callerIdentity: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DelegationCreationRequest"
      responses:
        "201":
          description: Delegation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DelegationResponse"
        "404":
          $ref: "#/components/responses/ManagementError"
        "409":
          $ref: "#/components/responses/ManagementError"
        "422":
          $ref: "#/components/responses/ManagementError"
  /gauth/mgmt/v1/mandates/{id}/delegation-chain:
    get:
      operationId: getDelegationChain
      tags: [delegations]
      summary: Get delegation chain
      description: Returns full ancestry chain from root to the specified mandate
      security:
        - callerIdentity: []
      parameters:
        - $ref: "#/components/parameters/MandateId"
      responses:
        "200":
          description: Delegation chain
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DelegationChainResponse"
        "404":
          $ref: "#/components/responses/ManagementError"
components:
  securitySchemes:
    callerIdentity:
      type: apiKey
      in: header
      name: X-Caller-Identity
      description: Caller identity for audit and authorization
  parameters:
    MandateId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    ManagementError:
      description: Management API error
      content:
        application/json:
          schema:
            type: object
            required:
              - error
            properties:
              error:
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: string
                  message:
                    type: string
  schemas:
...
    GovernanceProfile:
      type: string
      enum: [minimal, standard, strict, enterprise, behoerde]
    MandateStatus:
      type: string
      enum: [DRAFT, ACTIVE, SUSPENDED, EXPIRED, REVOKED, BUDGET_EXCEEDED, SUPERSEDED, DELETED]
    ApprovalMode:
      type: string
      enum: [autonomous, supervised, four-eyes]
    Phase:
      type: string
      enum: [plan, build, run]
    ShellMode:
      type: string
      enum: [any, denylist, allowlist]
...
    MandateScope:
      type: object
      required:
        - governance_profile
        - phase
      properties:
        governance_profile:
          $ref: "#/components/schemas/GovernanceProfile"
        phase:
          $ref: "#/components/schemas/Phase"
        core_verbs:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/ToolPolicy"
        platform_permissions:
          $ref: "#/components/schemas/PlatformPermissions"
        active_modules:
          type: array
          items:
            type: string
        allowed_paths:
          type: array
          items:
            type: string
        denied_paths:
          type: array
          items:
            type: string
        allowed_sectors:
          type: array
          items:
            type: string
        allowed_regions:
          type: array
          items:
            type: string
        allowed_transactions:
          type: array
          items:
            type: string
        transaction_matrix:
          type: object
          additionalProperties: true
        allowed_decisions:
          type: array
          items:
            type: string
...
    Parties:
      type: object
      required:
        - subject
        - customer_id
        - project_id
        - issued_by
      properties:
        subject:
          type: string
          minLength: 1
        customer_id:
          type: string
          minLength: 1
        project_id:
          type: string
          minLength: 1
        issued_by:
          type: string
          minLength: 1
        approval_chain:
          type: array
          items:
            type: string
    BudgetDetail:
      type: object
      required:
        - total_cents
        - remaining_cents
        - consumed_cents
        - utilization_percent
        - reserved_for_delegations_cents
      properties:
        total_cents:
          type: integer
        remaining_cents:
          type: integer
        consumed_cents:
          type: integer
        utilization_percent:
          type: number
        reserved_for_delegations_cents:
          type: integer
    MandateCreationRequest:
      type: object
      required:
        - parties
        - scope
        - requirements
      properties:
        parties:
          $ref: "#/components/schemas/Parties"
        scope:
          $ref: "#/components/schemas/MandateScope"
        requirements:
          $ref: "#/components/schemas/MandateRequirements"
...
    CreationResponse:
      type: object
      required:
        - mandate_id
        - status
        - governance_profile
        - scope_checksum
        - tool_permissions_hash
        - platform_permissions_hash
        - created_at
        - validation
      properties:
        mandate_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [DRAFT]
        governance_profile:
          $ref: "#/components/schemas/GovernanceProfile"
        scope_checksum:
          type: string
        tool_permissions_hash:
          type: string
        platform_permissions_hash:
          type: string
        created_at:
          type: string
          format: date-time
        validation:
          $ref: "#/components/schemas/ValidationResult"
    MandateResponse:
      type: object
      required:
        - mandate_id
        - status
        - governance_profile
        - phase
        - parties
        - scope
        - requirements
        - budget
        - scope_checksum
        - tool_permissions_hash
        - platform_permissions_hash
        - delegation_depth
        - parent_mandate_id
        - ttl_seconds
        - remaining_ttl_seconds
        - created_at
        - activated_at
        - expires_at
        - updated_at
      properties:
        mandate_id:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/MandateStatus"
        governance_profile:
          $ref: "#/components/schemas/GovernanceProfile"
        phase:
          $ref: "#/components/schemas/Phase"
        parties:
          $ref: "#/components/schemas/Parties"
        scope:
          $ref: "#/components/schemas/MandateScope"
        requirements:
          $ref: "#/components/schemas/MandateRequirements"
        budget:
          $ref: "#/components/schemas/BudgetDetail"
        scope_checksum:
          type: string
        tool_permissions_hash:
          type: string
        platform_permissions_hash:
          type: string
        delegation_depth:
          type: integer
        parent_mandate_id:
          type: string
          format: uuid
          nullable: true
        ttl_seconds:
          type: integer
        remaining_ttl_seconds:
          type: integer
        created_at:
          type: string
          format: date-time
        activated_at:
          type: string
          format: date-time
          nullable: true
        expires_at:
          type: string
          format: date-time
          nullable: true
        updated_at:
          type: string
          format: date-time
    PaginatedMandateResponse:
      type: object
      required:
        - items
        - next_cursor
        - total
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/MandateResponse"
        next_cursor:
          type: string
          nullable: true
        total:
          type: integer
    DeleteResponse:
      type: object
      required:
        - mandate_id
        - deleted
      properties:
        mandate_id:
          type: string
          format: uuid
        deleted:
          type: boolean
    AuditLogEntry:
      type: object
      required:
        - log_id
        - mandate_id
        - operation_type
        - caller_identity
        - detail
        - created_at
      properties:
        log_id:
          type: string
          format: uuid
        mandate_id:
          type: string
          format: uuid
        operation_type:
          type: string
        caller_identity:
          type: string
        detail:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
    ActivationResponse:
      type: object
      required:
        - mandate_id
        - status
        - activated_at
        - expires_at
      properties:
        mandate_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [ACTIVE]
        activated_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
    RevocationResponse:
      type: object
      required:
        - mandate_id
        - status
        - reason
        - revoked_at
      properties:
        mandate_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [REVOKED]
        reason:
          type: string
        revoked_at:
          type: string
          format: date-time
    SuspensionResponse:
      type: object
      required:
        - mandate_id
        - status
        - reason
        - suspended_at
      properties:
        mandate_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [SUSPENDED]
        reason:
          type: string
        suspended_at:
          type: string
          format: date-time
    ResumptionResponse:
      type: object
      required:
        - mandate_id
        - status
        - resumed_at
      properties:
        mandate_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [ACTIVE]
        resumed_at:
          type: string
          format: date-time
    BudgetIncreaseResponse:
      type: object
      required:
        - mandate_id
        - budget
        - additional_cents
        - increased_by
      properties:
        mandate_id:
          type: string
          format: uuid
        budget:
          $ref: "#/components/schemas/BudgetDetail"
        additional_cents:
          type: integer
        increased_by:
          type: string
    ConsumptionResponse:
      type: object
      required:
        - mandate_id
        - enforcement_request_id
        - amount_cents
        - budget
      properties:
        mandate_id:
          type: string
          format: uuid
        enforcement_request_id:
          type: string
        amount_cents:
          type: integer
        budget:
          $ref: "#/components/schemas/BudgetDetail"
    TTLExtensionResponse:
      type: object
      required:
        - mandate_id
        - ttl_seconds
        - expires_at
        - additional_seconds
        - extended_by
      properties:
        mandate_id:
          type: string
          format: uuid
        ttl_seconds:
          type: integer
        expires_at:
          type: string
          format: date-time
        additional_seconds:
          type: integer
        extended_by:
          type: string
    DelegationCreationRequest:
      type: object
      required:
        - parent_mandate_id
        - delegate_agent_id
        - budget_cents
        - ttl_seconds
      properties:
        parent_mandate_id:
          type: string
          minLength: 1
        delegate_agent_id:
          type: string
          minLength: 1
        scope_restriction:
          type: object
          additionalProperties: true
          default: {}
        budget_cents:
          type: integer
          minimum: 0
        ttl_seconds:
          type: integer
          minimum: 60
    DelegationResponse:
      type: object
      required:
        - mandate_id
        - parent_mandate_id
        - delegate_agent_id
        - delegation_depth
        - budget
        - scope
        - created_at
        - expires_at
      properties:
        mandate_id:
          type: string
          format: uuid
        parent_mandate_id:
          type: string
          format: uuid
        delegate_agent_id:
          type: string
        delegation_depth:
          type: integer
        budget:
          $ref: "#/components/schemas/BudgetDetail"
        scope:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
...
    DelegationChainResponse:
      type: object
      required:
        - mandate_id
        - chain
      properties:
        mandate_id:
          type: string
          format: uuid
        chain:
          type: array
          items:
            $ref: "#/components/schemas/DelegationChainEntry"
    ProfileInfo:
      type: object
      required:
        - name
        - description
        - registration_context
      properties:
        name:
          $ref: "#/components/schemas/GovernanceProfile"
        description:
          type: string
        registration_context:
          type: string
    ProfileCeilings:
      type: object
      required:
        - name
      properties:
        name:
          $ref: "#/components/schemas/GovernanceProfile"
        deployment_targets:
          type: array
          items:
            type: string
        auto_deploy:
          type: boolean
        db_write:
          type: boolean
        db_migration:
          type: boolean
        db_production:
          type: boolean
        shell_mode:
          $ref: "#/components/schemas/ShellMode"
        packages_audited_only:
          type: boolean
        secrets_read:
          type: boolean
        secrets_create:
          type: boolean
        agent_delegation:
          type: boolean
        max_delegation_depth:
          type: integer
        min_approval_mode:
          $ref: "#/components/schemas/ApprovalMode"
        max_session_duration_minutes:
          type: integer
          nullable: true
        max_tool_calls:
          type: integer
          nullable: true
        max_lines_per_commit:
          type: integer
          nullable: true
        description:
          type: string
        registration_context:
          type: string
...